About: Update premise
I can't promise to write daily. It would be a great habit to have, but
in reality I find it very stressful to force myself to write.

Most often the free time I get ends up way late in the evening after a
hard day of work and only a semi-relaxing evening as a parent. A lot of
times I honestly just want to empty my brain with TV, reading or other
non-productive stuff. And that's totally fine.

Being creative takes time and I accept that right now I just can't find
the time to do so very often.