Add an imprint
I hate this! If I were strictly following regulations I would even have
to add my private address.

As it is, my name is all over the page anyway, and it is easy enough to
get my mail from any of the linked services (GitHub, GitLab). Where I
live is also no big secret and the city is big enough that I'm not that
easy to find. Hopefully.