Hugo: Add licensing information to *almost* all posts
The only exception is "Phoebe", because I've participated in a contest
with this story and found conflicting information concerning the
license.

I've already emailed the organizer and will update the license as soon
as I'm sure. It's between CC BY-NC-ND and CC BY-NC-SA. Unless I know for
sure that it's ND I don't see any reason to include a different license
in the template.

For anyone interested: the difference between SA (Share Alike) and ND
(No Derivatives) is that SA allows people to adapt or remix the content
and publish it again.