Hugo: Add content warning for specific tags
This adds a generic content warning to stories that are tagged as
"horror" or "brutal".

I feel like the specific content warnings always gave away a part of the
story, but I know for a fact that some of my readers want to skip
certain stories that might contain disturbing elements.