Add new mindfulness posts
I decided to start using drafts and not separating every commit by post.
It's getting ridiculous and if I ever go back to edit a post, the diff
is still visible in a new commit.

I'm also thinking of scheduling posts with a scheduled pipeline, but I'm
not yet fully committed to the idea. It'd be nice if Hugo could do this
out of the box, but I realize that it's just generating a **static**
site, so there can't be dynamically published posts on it.