Hugo: Add page for RSS links and move imprint
The imprint shouldn't be a top level section. Nobody cares about that (I
hope).

But what people do care about is a way to subscribe only to certain
parts of the site. Hugo generates the RSS feeds automatically, which is
nice and it was pretty easy to figure out the address. All I had to do
was add a new shortcode to generate the HTML list and voilà.